requirements.txt
# Requirements pour le projet Mesures DNS
# Installation: pip install -r requirements.txt

# === DNS et Reseau ===
dnspython>=2.4.0
ripe.atlas.cousteau>=1.4.0
ripe.atlas.sagan>=1.3.0
ripe.atlas.tools>=3.0.0
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
ipaddress
netaddr>=0.8.0
tldextract>=5.1.0
publicsuffixlist>=0.10.0

# === Analyse de donnees ===
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# === Visualisation ===
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
bokeh>=3.3.0
altair>=5.1.0
folium>=0.15.0
geopandas>=0.14.0

# === PDF ===
PyPDF2>=3.0.0
pypdf>=3.16.0
pdfplumber>=0.10.0
pdf2image>=1.16.0
pikepdf>=8.7.0
reportlab>=4.0.0
pdfminer.six>=20221105
pymupdf>=1.23.0

# === Documents Office ===
python-docx>=1.1.0
python-pptx>=0.6.0
openpyxl>=3.1.0
xlrd>=2.0.0
xlsxwriter>=3.1.0
odfpy>=1.4.0
mammoth>=1.6.0

# === Images ===
Pillow>=10.1.0
opencv-python-headless>=4.8.0
imageio>=2.33.0
scikit-image>=0.22.0
svglib>=1.5.0
cairosvg>=2.7.0

# === LaTeX et Markdown ===
pylatex>=1.4.0
markdown>=3.5.0
markdown2>=2.4.0
mistune>=3.0.0

# === Jupyter ===
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0
ipywidgets>=8.1.0
nbconvert>=7.11.0

# === Base de donnees ===
sqlalchemy>=2.0.0
sqlite-utils>=3.35.0
tinydb>=4.8.0
h5py>=3.10.0
pyarrow>=14.0.0
fastparquet>=2023.10.0

# === Utilitaires ===
python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.66.0
rich>=13.7.0
click>=8.1.0
typer>=0.9.0
pyyaml>=6.0.0
toml>=0.10.0
python-dotenv>=1.0.0
loguru>=0.7.0
tabulate>=0.9.0
prettytable>=3.9.0

# === Web scraping ===
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# === Serialisation ===
orjson>=3.9.0
ujson>=5.9.0
msgpack>=1.0.0
cbor2>=5.5.0

# === Tests ===
pytest>=7.4.0
pytest-asyncio>=0.21.0
